move yahp test to hparams folder
