+++ tests/hparams/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2022 MosaicML Composer authors
+# SPDX-License-Identifier: Apache-2.0
