+++ tests/hparams/test_algorithm_registry.py
