+++ tests/hparams/test_dataset_registry.py
