+++ tests/hparams/test_optimizer_registry.py
@@ -8,8 +8,8 @@ import pytest
 import composer.optim.optimizer_hparams_registry
 from composer.optim.optimizer_hparams_registry import OptimizerHparams, optimizer_registry
 from tests.common import get_module_subclasses
-from tests.common.hparams import assert_in_registry, construct_from_yaml
 from tests.common.models import SimpleModel
+from tests.hparams.common import assert_in_registry, construct_from_yaml
 
 optimizer_hparam_classes = get_module_subclasses(composer.optim.optimizer_hparams_registry, OptimizerHparams)
